requirements.txt
django==4.1.7
asgiref==3.5.2
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-dbbackup==4.0.2
django-dynamic-preferences==1.14.0
django-anymail[mailgun]==8.6
django-anymail[sendgrid]==8.6
psycopg2-binary==2.9.5
redis==4.3.5
pyjwt==2.6.0
pytz==2022.6
requests==2.28.1
twilio==7.15.4
django-phonenumber-field==7.0.0
phonenumbers==8.13.1
pillow==9.3.0
django-storages==1.13.1
django-sendsms==0.5
boto3==1.26.22
awscli==1.27.22
pyotp==2.7.0
cerberus==1.3.4
django-model-utils==4.3.1
mysql-connector-python==8.0.31
stripe==5.0.0
huey==2.4.4
rollbar==0.16.3
django-ckeditor-5==0.2.1
semver==2.13.0
django-cal==0.4.2
django-allauth==0.52.0 # https://github.com/pennersr/django-allauth
drf-access-policy==1.3.0 # https://github.com/rsinger86/drf-access-policy
